Separate list_datatable JS into fine-grained modules
No doubt these modules could themselves benefit from some standardization and refactoring; the point here is partly to prototype a generic approach to datatable logic which could cover other uses of datables too.
