+++ app/javascript/packs/list_datatable/columns/basic_column.js
@@ -1,0 +1,9 @@
+/** A generic column template with some sensible defaults */
+export function BasicColumn(dataname, name){
+  return {
+    data: dataname,
+    name: name || dataname,
+    visible: false,
+    searchable: true
+  }
+}
