+++ app/javascript/packs/list_datatable/columns/context_column.js
@@ -1,0 +1,11 @@
+export function ContextColumn(config) {
+  if ( config['include_context_col'] ){
+    return {
+      data: 'context',
+      name: 'context',
+      width: '30%',
+      sortable: false,
+      className: 'context'
+    }
+  }
+}
