+++ app/javascript/packs/list_datatable/columns/id_column.js
@@ -1,0 +1,6 @@
+export const IdColumn = {
+  data: 'id',
+  name: 'id',
+  visible: false,
+  searchable: false
+}
