+++ app/javascript/packs/list_datatable/columns/master_search_column.js
@@ -1,0 +1,12 @@
+export function MasterSearchColumn(){
+  return {
+    data: 'master_search',
+    name: 'master_search',
+    visible: false,
+    searchable: true,
+    render: function(data, type, row) {
+      var keys = ['name', 'blurb', 'types', 'industries']
+        return keys.map(function(key) { return row[key] }).join('  ');
+    }
+  }
+}
